LSM6DSO32: removed a lot of duplicated code
With dedicated converter methods, the LSM6DSO32 can do its conversion for
the bigger ranges there, so duplicated code in _read() can be removed.
Also, getAccelRange() and setAccelRange() are exactly the same as in
the base class, the difference is only the enum type. Replaced these
methods with a call to the baseclass with the casted range.