added a selector to _read()
With this selector (which defaults to the behaviour as before this
commit), the calling code can specify which kind data to transfer from the
IMU. All the data transfer code is centralized in _read() now, the various
getEvent() methods only transfer the data they need. I removed some hard
coded conversions (all read outs now respect the set ranges) and also added
a method to read out the accelerometer and gyroscope data at the same time.