github-actions: do the more expensive tests last
In my experience, the clang-format and doxygen tests failed the most
in my PRs. These are also the least expensive tests, they execute in
the range of seconds, but are last checked. You can reasonally expect
that the code compiles, but documentation and code formating are hit
and miss, especially when different people have different editors and
workflows. This saves a lot of execution time on github and gives the
users more immediate feedback: 10 out of 13 minutes are the compilation
on different platforms you have to wait for until the it fails on a
trivial formating or documentation error.